--- browser.js
@@ -1,1 +1,0 @@
-module.exports = require('./inject')(require('random-access-web'), require('lru_cache').LRUCache)
